Migrate FileServiceTests
